Revert "Refactoring 3"
