+++ SwiftCards/JoinerViewController.swift
@@ -1,15 +1,22 @@
 import UIKit
 
 class JoinerViewController: UIViewController {
+
+    var activityIndicator: UIActivityIndicatorView = UIActivityIndicatorView()
+
     override func viewDidLoad() {
         super.viewDidLoad()
-        setupActivityIndicator()
-    }
-    func setupActivityIndicator() {
-        let activityIndicator = UIActivityIndicatorView()
         activityIndicator.center = self.view.center
         activityIndicator.activityIndicatorViewStyle = UIActivityIndicatorViewStyle.gray
         view.addSubview(activityIndicator)
         activityIndicator.startAnimating()
     }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+    }
+
+    override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
+    }
+
 }
